backend/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Cors==3.0.10
requests==2.26.0
python-dotenv==0.19.1
nltk==3.6.5
textblob==0.15.3
vaderSentiment==3.3.2
tweepy==4.4.0
praw==7.5.0
pandas==1.3.4
plotly==5.5.0
newsapi-python==0.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
